move-variation analysis section

+++ include/global.h
@@ -6,8 +6,8 @@
 extern const double g_chessboardWidth;
 extern const double g_chessboardHeight;
 extern const double g_squareSize;
-extern const int g_pixel_dx;
-extern const int g_pixel_dy;
+extern const float g_pixel_dx;
+extern const float g_pixel_dy;
 extern const std::string board_[8][8];
 
 #endif // GLOBAL_H
